+++ cdc-office-tables/src/main/java/cdc/office/tables/diff/KeyedTableDiff.java
@@ -17,30 +17,50 @@ import cdc.util.lang.InvalidDataException;
 import cdc.util.lang.UnexpectedValueException;
 
 /**
- * Class used to compare two lists of rows identified by a key set.
+ * Class used to compare two lists of rows identified by a set of keys.
  *
  * @author Damien Carbonne
- *
  */
 public class KeyedTableDiff {
-    // private static final Logger LOGGER = LogManager.getLogger(KeyedTableDiff.class);
+    /** The left header. */
     private final Header leftHeader;
+    /** The right header. */
     private final Header rightHeader;
+    /** The names of key columns. */
     private final List<String> keyNames;
+    /** Indices in left header of key columns. */
     private final int[] leftKeyColumns;
+    /** Indices in right header of key columns. */
     private final int[] rightKeyColumns;
+    /**
+     * The list of all keys.
+     * <p>
+     * Some may be left or right only, some may be common.
+     */
     private final List<CTupleN<String>> keys = new ArrayList<>();
+    /** Maps from keys to row differences. */
     private final Map<CTupleN<String>, RowDiff> diffs = new HashMap<>();
+    /** The synthesis of differences. */
+    private final Synthesis synthesis = new Synthesis();
 
+    /**
+     * Creates a KeyedTableDiff.
+     *
+     * @param leftHeader The left header.
+     * @param leftRows The left rows.
+     * @param rightHeader the right header.
+     * @param rightRows The right rows.
+     * @param keyNames The key names.
+     */
     public KeyedTableDiff(Header leftHeader,
-                          List<Row> leftData,
+                          List<Row> leftRows,
                           Header rightHeader,
-                          List<Row> rightData,
+                          List<Row> rightRows,
                           String... keyNames) {
         Checks.isNotNull(leftHeader, "leftHeader");
-        Checks.isNotNull(leftData, "leftData");
+        Checks.isNotNull(leftRows, "leftRows");
         Checks.isNotNull(rightHeader, "rightHeader");
-        Checks.isNotNull(rightData, "rightData");
+        Checks.isNotNull(rightRows, "rightRows");
         Checks.isNotNullOrEmpty(keyNames, "keys");
 
         this.leftHeader = leftHeader;
@@ -50,10 +70,10 @@ public class KeyedTableDiff {
         this.leftKeyColumns = buildKeyColumns(leftHeader, keyNames);
         this.rightKeyColumns = buildKeyColumns(rightHeader, keyNames);
 
-        // Map from tuples to left rows
+        // Map from left keys to left rows
         final Map<CTupleN<String>, Row> leftMap = new HashMap<>();
-        for (int number = 0; number < leftData.size(); number++) {
-            final Row row = leftData.get(number);
+        for (int number = 0; number < leftRows.size(); number++) {
+            final Row row = leftRows.get(number);
             final CTupleN<String> key = getKey(Side.LEFT, row, number);
             if (leftMap.containsKey(key)) {
                 throw new InvalidDataException("Duplicate key " + key + locate(Side.LEFT, row, number));
@@ -62,8 +82,8 @@ public class KeyedTableDiff {
         }
 
         // Compare right rows to left ones
-        for (int number = 0; number < rightData.size(); number++) {
-            final Row right = rightData.get(number);
+        for (int number = 0; number < rightRows.size(); number++) {
+            final Row right = rightRows.get(number);
             final CTupleN<String> key = getKey(Side.RIGHT, right, number);
             final Row left = leftMap.getOrDefault(key, Row.EMPTY);
             if (diffs.containsKey(key)) {
@@ -71,52 +91,78 @@ public class KeyedTableDiff {
             }
 
             final RowDiff diff = new RowDiff(leftHeader, left, rightHeader, right);
-            // LOGGER.info("{} -> {}", key, diff);
             diffs.put(key, diff);
             keys.add(key);
         }
 
         // Add all left rows that are not in right rows
-        for (int number = 0; number < leftData.size(); number++) {
-            final Row left = leftData.get(number);
+        for (int number = 0; number < leftRows.size(); number++) {
+            final Row left = leftRows.get(number);
             final CTupleN<String> key = getKey(Side.LEFT, left, number);
             if (!diffs.containsKey(key)) {
                 diffs.put(key, new RowDiff(leftHeader, left, rightHeader, Row.EMPTY));
                 keys.add(key);
             }
         }
+
+        this.synthesis.compute(this);
     }
 
+    /**
+     * Creates a KeyedTableDiff.
+     *
+     * @param leftHeader The left header.
+     * @param leftRows The left rows.
+     * @param rightHeader the right header.
+     * @param rightRows The right rows.
+     * @param keyNames The key names.
+     */
     public KeyedTableDiff(Header leftHeader,
-                          List<Row> leftData,
+                          List<Row> leftRows,
                           Header rightHeader,
-                          List<Row> rightData,
+                          List<Row> rightRows,
                           List<String> keyNames) {
         this(leftHeader,
-             leftData,
+             leftRows,
              rightHeader,
-             rightData,
+             rightRows,
              keyNames.toArray(new String[keyNames.size()]));
     }
 
+    /**
+     * Creates a KeyedTableDiff.
+     *
+     * @param header The common header.
+     * @param leftRows The left rows.
+     * @param rightRows The right rows.
+     * @param keyNames The key names.
+     */
     public KeyedTableDiff(Header header,
-                          List<Row> leftData,
-                          List<Row> rightData,
+                          List<Row> leftRows,
+                          List<Row> rightRows,
                           String... keyNames) {
         this(header,
-             leftData,
+             leftRows,
              header,
-             rightData,
+             rightRows,
              keyNames);
     }
 
+    /**
+     * Creates a KeyedTableDiff.
+     *
+     * @param header The common header.
+     * @param leftRows The left rows.
+     * @param rightRows The right rows.
+     * @param keyNames The key names.
+     */
     public KeyedTableDiff(Header header,
-                          List<Row> leftData,
-                          List<Row> rightData,
+                          List<Row> leftRows,
+                          List<Row> rightRows,
                           List<String> keyNames) {
         this(header,
-             leftData,
-             rightData,
+             leftRows,
+             rightRows,
              keyNames.toArray(new String[keyNames.size()]));
     }
 
@@ -176,67 +222,145 @@ public class KeyedTableDiff {
         return diffs.values();
     }
 
-    /**
-     * Prints a synthesis of this diff.
-     *
-     * @param out The PrintStream.
-     */
-    public void printSynthesis(PrintStream out) {
-        int addedLines = 0;
-        int removedLines = 0;
-        int unchangedLines = 0;
-        int changedLines = 0;
-        int addedCells = 0;
-        int removedCells = 0;
-        int unchangedCells = 0;
-        int changedCells = 0;
+    public Synthesis getSynthesis() {
+        return synthesis;
+    }
 
-        for (final RowDiff rdiff : getDiffs()) {
-            switch (rdiff.getKind()) {
-            case ADDED:
-                addedLines++;
-                break;
-            case CHANGED:
-                changedLines++;
-                for (final CellDiff cdiff : rdiff.getDiffs()) {
-                    switch (cdiff.getKind()) {
-                    case ADDED:
-                        addedCells++;
-                        break;
-                    case CHANGED:
-                        changedCells++;
-                        break;
-                    case REMOVED:
-                        removedCells++;
-                        break;
-                    case SAME:
-                    case NULL:
-                        unchangedCells++;
-                        break;
-                    default:
-                        throw new UnexpectedValueException(cdiff.getKind());
-                    }
+    /**
+     * Synthesis of differences (for statistics).
+     * <p>
+     * Numbers are computed for:
+     * <ul>
+     * <li>lines
+     * <li>cells
+     * <li>columns.
+     * </ul>
+     *
+     * @author Damien Carbonne
+     */
+    public static class Synthesis {
+        public enum Action {
+            ADDED,
+            REMOVED,
+            CHANGED,
+            UNCHANGED
+        }
+
+        /** Counts of lines. */
+        private final int[] lines = new int[Action.values().length];
+        /** Counts of cell. */
+        private final int[] cells = new int[Action.values().length];
+        /** Counts of cells by column. */
+        private final Map<String, int[]> columnToCells = new HashMap<>();
+        /** The list of all (left and right) column names. */
+        private final List<String> columnNames = new ArrayList<>();
+
+        private void compute(KeyedTableDiff diffs) {
+            // retrieve all column names.
+            this.columnNames.addAll(diffs.leftHeader.getNames());
+            for (final String name : diffs.rightHeader.getNames()) {
+                if (!columnNames.contains(name)) {
+                    columnNames.add(name);
                 }
-                break;
-            case REMOVED:
-                removedLines++;
-                break;
-            case SAME:
-                unchangedLines++;
-                break;
-            default:
-                throw new UnexpectedValueException(rdiff.getKind());
+            }
+            // Initialize map
+            for (final String name : columnNames) {
+                columnToCells.put(name, new int[Action.values().length]);
+            }
+
+            for (final RowDiff rdiff : diffs.getDiffs()) {
+                switch (rdiff.getKind()) {
+                case ADDED:
+                    lines[Action.ADDED.ordinal()]++;
+                    cells[Action.ADDED.ordinal()] += diffs.rightHeader.size();
+                    for (final String name : diffs.rightHeader.getNames()) {
+                        final int[] counts = columnToCells.get(name);
+                        counts[Action.ADDED.ordinal()]++;
+                    }
+                    break;
+                case CHANGED:
+                    lines[Action.CHANGED.ordinal()]++;
+                    for (final LocalizedCellDiff lcdiff : rdiff.getDiffs()) {
+                        final int[] counts = columnToCells.get(lcdiff.getKey());
+                        switch (lcdiff.getDiff().getKind()) {
+                        case ADDED:
+                            cells[Action.ADDED.ordinal()]++;
+                            counts[Action.ADDED.ordinal()]++;
+                            break;
+                        case CHANGED:
+                            cells[Action.CHANGED.ordinal()]++;
+                            counts[Action.CHANGED.ordinal()]++;
+                            break;
+                        case REMOVED:
+                            cells[Action.REMOVED.ordinal()]++;
+                            counts[Action.REMOVED.ordinal()]++;
+                            break;
+                        case SAME:
+                        case NULL:
+                            cells[Action.UNCHANGED.ordinal()]++;
+                            counts[Action.UNCHANGED.ordinal()]++;
+                            break;
+                        default:
+                            throw new UnexpectedValueException(lcdiff.getDiff().getKind());
+                        }
+                    }
+                    break;
+                case REMOVED:
+                    lines[Action.REMOVED.ordinal()]++;
+                    cells[Action.REMOVED.ordinal()] += diffs.leftHeader.size();
+                    for (final String name : diffs.leftHeader.getNames()) {
+                        final int[] counts = columnToCells.get(name);
+                        counts[Action.REMOVED.ordinal()]++;
+                    }
+                    break;
+                case SAME:
+                    lines[Action.UNCHANGED.ordinal()]++;
+                    // left and right header have the same size
+                    cells[Action.UNCHANGED.ordinal()] += diffs.leftHeader.size();
+                    for (final String name : diffs.leftHeader.getNames()) {
+                        final int[] counts = columnToCells.get(name);
+                        counts[Action.UNCHANGED.ordinal()]++;
+                    }
+                    break;
+                default:
+                    throw new UnexpectedValueException(rdiff.getKind());
+                }
             }
         }
-        out.println("Lines");
-        out.println("   Added:     " + addedLines);
-        out.println("   Removed:   " + removedLines);
-        out.println("   Changed:   " + changedLines);
-        out.println("   Unchanged: " + unchangedLines);
-        out.println("Cells");
-        out.println("   Added:     " + addedCells);
-        out.println("   Removed:   " + removedCells);
-        out.println("   Changed:   " + changedCells);
-        out.println("   Unchanged: " + unchangedCells);
+
+        public int getLinesCount(Action action) {
+            return lines[action.ordinal()];
+        }
+
+        public int getCellsCount(Action action) {
+            return cells[action.ordinal()];
+        }
+
+        public List<String> getColumnNames() {
+            return columnNames;
+        }
+
+        public int getColumnCellsCount(String columnName,
+                                       Action action) {
+            return columnToCells.get(columnName)[action.ordinal()];
+        }
+
+        public void print(PrintStream out) {
+            final String format = "   %-10s %d%n";
+            out.println("Lines");
+            for (final Synthesis.Action action : Synthesis.Action.values()) {
+                out.printf(format, action + ":", getLinesCount(action));
+            }
+            out.println("Cells");
+            for (final Synthesis.Action action : Synthesis.Action.values()) {
+                out.printf(format, action + ":", getCellsCount(action));
+            }
+            for (final String name : getColumnNames()) {
+                out.println("[" + name + "]");
+                for (final Synthesis.Action action : Synthesis.Action.values()) {
+                    out.printf(format, action + ":", getColumnCellsCount(name, action));
+                }
+            }
+        }
     }
 }